requirements.txt
asyncpraw
python-dotenv
discord